requirements.txt
numpy
pandas
scikit-learn
xgboost
mlflow
python-dotenv
fastapi
uvicorn[standard]
pyyaml
dvc
